fix: DX-2713 - 'Your dataset is here' tooltip after uploading a dataset is not pointing at the right dataset and goes to any dataset the cursor is on
